src/test/java/katas/ReverseWordsTest.java
package katas;

import org.junit.Test;

import static org.junit.Assert.*;

public class ReverseWordsTest {

    @Test
    public void reverseWords() {
        assertEquals("ehT kciuq nworb xof spmuj revo eht yzal .god", ReverseWords.reverseWords("The quick brown fox jumps over " +
                "the lazy dog."));
        assertEquals("elppa", ReverseWords.reverseWords("apple"));
        assertEquals("a b c d", ReverseWords.reverseWords("a b c d"));
        assertEquals("elbuod  decaps  sdrow", ReverseWords.reverseWords("double  spaced  words"));
    }

    @Test
    public void getKataName() {
        assertEquals("https://www.codewars.com/kata/5259b20d6021e9e14c0010d4/train/java", ReverseWords.getKataName());
    }
}